Datasets refactoring
Renamed dataset storage location
