+++ backend/dataall/modules/datasets/api/storage_location/__init__.py
@@ -1,4 +1,4 @@
-from dataall.modules.datasets.api.DatasetStorageLocation import (
+from dataall.modules.datasets.api.storage_location import (
     input_types,
     mutations,
     queries,
