+++ backend/dataall/modules/datasets/api/storage_location/queries.py
@@ -1,5 +1,5 @@
 from dataall.api import gql
-from dataall.modules.datasets.api.DatasetStorageLocation.resolvers import get_storage_location
+from dataall.modules.datasets.api.storage_location.resolvers import get_storage_location
 
 getDatasetStorageLocation = gql.QueryField(
     name='getDatasetStorageLocation',
