+++ backend/dataall/modules/datasets/api/storage_location/schema.py
@@ -1,5 +1,5 @@
 from dataall.api import gql
-from dataall.modules.datasets.api.DatasetStorageLocation.resolvers import (
+from dataall.modules.datasets.api.storage_location.resolvers import (
     resolve_glossary_terms,
     resolve_dataset
 )
